(android) Update the payment screens for liquidity purchases
With the new on-the-fly mechanism, the automated liquidity
operation necesssary to receive some payments are stored as
specific payments. That is, the incoming payments related
to this operation do not have a fee, instead there's a
payment row in the payments history specific to the liquidity
purchase. The operation also has its own payment details page
just like the (already existing) manual liquidity purchase.

The splash screen file has been split into several files, one
for each payment types. This makes the code more flexible.

The liquidity policy objects set in the channels management
screen have been updated to match the NodeParamsManager value.

This commit is a WIP, especially for wording and localisation.
