+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/settings/NotificationsView.kt
@@ -268,14 +268,14 @@ private fun PaymentNotification(
                         notification.fee.toPrettyString(btcUnit, withUnit = true),
                         notification.maxAbsoluteFee.toPrettyString(btcUnit, withUnit = true),
                     )
-
                     is Notification.OverRelativeFee -> stringResource(
                         id = R.string.inappnotif_payment_rejected_over_relative,
                         notification.fee.toPrettyString(btcUnit, withUnit = true),
                         DecimalFormat("0.##").format(notification.maxRelativeFeeBasisPoints.toDouble() / 100),
                     )
-
-                    is Notification.ChannelsInitializing -> stringResource(id = R.string.inappnotif_payment_rejected_channel_initializing)
+                    is Notification.GenericError -> "An error has occurred. Please try again."
+                    is Notification.ChannelFundingInProgress -> "A funding is in progress. Try again later."
+                    is Notification.MissingOffChainAmountTooLow -> "The amount is too low."
                 },
                 bottomText = when (notification) {
                     is Notification.OverAbsoluteFee, is Notification.OverRelativeFee, is Notification.FeePolicyDisabled -> {
