+++ phoenix-android/src/main/kotlin/fr/acinq/phoenix/android/settings/walletinfo/SwapInWalletInfo.kt
@@ -237,7 +237,9 @@ private fun ReadyForSwapView(
                         DecimalFormat("0.##").format(lastSwapFailedNotification.maxRelativeFeeBasisPoints.toDouble() / 100),
                     )
                     is Notification.FeePolicyDisabled -> stringResource(id = R.string.walletinfo_onchain_swapin_last_attempt_disabled)
-                    is Notification.ChannelsInitializing -> stringResource(id = R.string.walletinfo_onchain_swapin_last_attempt_channels_init)
+                    is Notification.ChannelFundingInProgress -> stringResource(id = R.string.walletinfo_onchain_swapin_last_attempt_funding_in_progress)
+                    is Notification.MissingOffChainAmountTooLow -> stringResource(id = R.string.walletinfo_onchain_swapin_last_attempt_amount_too_low)
+                    is Notification.GenericError -> stringResource(id = R.string.walletinfo_onchain_swapin_last_attempt_generic)
                 },
             )
         }
