+++ phoenix-android/src/main/res/values-b+es+419/strings.xml
@@ -330,7 +330,7 @@
     <string name="paymentdetails_lnurlpay_action_aes_label">Mensaje</string>
     <string name="paymentdetails_lnurlpay_action_aes_decrypting">Desencriptando mensaje…</string>
 
-    <string name="paymentdetails_desc_label">Desc.</string>
+    <string name="paymentdetails_desc_label">Descripción</string>
     <string name="paymentdetails_destination_label">Enviado a</string>
     <string name="paymentdetails_destination_cpfp_value">Mineros de Bitcoin</string>
     <string name="paymentdetails_fees_label">Comisiones</string>
