+++ phoenix-android/src/main/res/values-cs/important_strings.xml
@@ -44,7 +44,6 @@
     <string name="notif_rejected_over_absolute_timeout">Poplatek činil %1$s, ale váš maximální poplatek byl nastaven na %2$s. Platnost této zálohy vyprší %3$s.</string>
     <string name="notif_rejected_over_relative">Poplatek činil %1$s, což je více než %2$s%% z přijaté částky. Můžete to upravit v nastavení.</string>
     <string name="notif_rejected_over_relative_timeout">Poplatek činil %1$s, což je více než %2$s%% z přijaté částky. Platnost této zálohy vyprší %3$s.</string>
-    <string name="notif_rejected_channels_initializing">Vaše kanály se inicializovaly a nemohly tuto platbu přijmout. Zkuste to později.</string>
 
     <string name="notif_watcher_revoked_commit_title">Spusťte prosím Phoenix</string>
     <string name="notif_watcher_revoked_commit_message">Některý z vašich kanálů mohl být uzavřen.</string>
