+++ phoenix-android/src/main/res/values-de/strings.xml
@@ -337,7 +337,7 @@
     <string name="paymentdetails_lnurlpay_action_aes_label">Nachricht</string>
     <string name="paymentdetails_lnurlpay_action_aes_decrypting">Nachricht entschlüsseln..</string>
 
-    <string name="paymentdetails_desc_label">Desc.</string>
+    <string name="paymentdetails_desc_label">Beschreibung</string>
     <string name="paymentdetails_destination_label">Gesendet an</string>
     <string name="paymentdetails_destination_cpfp_value">Bitcoin-Miner</string>
     <string name="paymentdetails_fees_label">Gebühren</string>
@@ -382,7 +382,6 @@
     <string name="paymentdetails_dualswapin_tx_value">- #%1$s:</string>
 
     <string name="paymentdetails_liquidity_amount_label">Geforderter Betrag</string>
-    <string name="paymentdetails_liquidity_signature_label">Unterschrift</string>
 
     <string name="paymentdetails_closing_type_label">Schließungs-Typ</string>
     <string name="paymentdetails_closing_type_mutual">Einvernehmlich</string>
