+++ phoenix-android/src/main/res/values-fr/strings.xml
@@ -380,7 +380,6 @@
     <string name="paymentdetails_dualswapin_tx_value">- #%1$s:</string>
 
     <string name="paymentdetails_liquidity_amount_label">Liquidité demandée</string>
-    <string name="paymentdetails_liquidity_signature_label">Signature</string>
 
     <string name="paymentdetails_closing_type_label">Type de clôture</string>
     <string name="paymentdetails_closing_type_mutual">Mutuelle</string>
