+++ phoenix-android/src/main/res/values/strings.xml
@@ -349,6 +349,7 @@
     <string name="paymentdetails_error">Could not find payment details</string>
 
     <string name="paymentdetails_status_inbound_liquidity_success"><b>LIQUIDITY ADDED</b> %1$s</string>
+    <string name="paymentdetails_status_inbound_liquidity_auto_success"><b>CHANNEL RESIZED</b> %1$s</string>
     <string name="paymentdetails_status_channelclose_confirmed"><b>COMPLETE</b> %1$s</string>
     <string name="paymentdetails_status_sent_successful"><b>SENT</b> %1$s</string>
     <string name="paymentdetails_status_sent_pending">Pending…</string>
@@ -378,7 +379,7 @@
     <string name="paymentdetails_lnurlpay_action_aes_label">Message</string>
     <string name="paymentdetails_lnurlpay_action_aes_decrypting">Decrypting message…</string>
 
-    <string name="paymentdetails_desc_label">Desc.</string>
+    <string name="paymentdetails_desc_label">Description</string>
     <string name="paymentdetails_destination_label">Sent to</string>
     <string name="paymentdetails_destination_cpfp_value">Bitcoin miners</string>
     <string name="paymentdetails_fees_label">Fees</string>
@@ -425,7 +426,6 @@
     <string name="paymentdetails_dualswapin_tx_value">- #%1$s:</string>
 
     <string name="paymentdetails_liquidity_amount_label">Amount requested</string>
-    <string name="paymentdetails_liquidity_signature_label">Signature</string>
 
     <string name="paymentdetails_closing_type_label">Closing type</string>
     <string name="paymentdetails_closing_type_mutual">Mutual</string>
@@ -459,6 +459,7 @@
     <string name="paymentdetails_received_with_splicein">Splice-in (adding to existing channel)</string>
     <string name="paymentdetails_received_with_channel">New channel (automatically created)</string>
     <string name="paymentdetails_received_with_lightning">Lightning payment</string>
+    <string name="paymentdetails_received_with_fee_credit">Fee credit</string>
     <string name="paymentdetails_channel_id_label">Channel id</string>
     <string name="paymentdetails_tx_id_label">Transaction</string>
 
@@ -470,6 +471,8 @@
     <string name="paymentdetails_invoice_requested_label">Amount requested</string>
     <string name="paymentdetails_amount_sent_label">Amount sent (fees included)</string>
     <string name="paymentdetails_amount_received_label">Amount received</string>
+    <string name="paymentdetails_amount_fee_credit_label">Fee credit accrued</string>
+    <string name="paymentdetails_amount_added_to_fee_credit_label">Amount added to fee credit</string>
     <string name="paymentdetails_amount_rate_now">≈ %1$s (now)</string>
     <string name="paymentdetails_amount_rate_then">≈ %1$s (then)</string>
